Rename PopupNotificationsManager to NotificationsManager
