+++ Habitica/src/main/java/com/habitrpg/android/habitica/data/implementation/ApiClientImpl.kt
@@ -14,7 +14,7 @@ import com.habitrpg.android.habitica.api.HostConfig
 import com.habitrpg.android.habitica.api.Server
 import com.habitrpg.android.habitica.data.ApiClient
 import com.habitrpg.android.habitica.events.ShowConnectionProblemEvent
-import com.habitrpg.android.habitica.helpers.PopupNotificationsManager
+import com.habitrpg.android.habitica.helpers.NotificationsManager
 import com.habitrpg.android.habitica.models.*
 import com.habitrpg.android.habitica.models.auth.UserAuth
 import com.habitrpg.android.habitica.models.auth.UserAuthResponse
@@ -61,7 +61,7 @@ import javax.net.ssl.SSLException
 
 class ApiClientImpl//private OnHabitsAPIResult mResultListener;
 //private HostConfig mConfig;
-(private val gsonConverter: GsonConverterFactory, override val hostConfig: HostConfig, private val crashlyticsProxy: CrashlyticsProxy, private val popupNotificationsManager: PopupNotificationsManager, private val context: Context) : Consumer<Throwable>, ApiClient {
+(private val gsonConverter: GsonConverterFactory, override val hostConfig: HostConfig, private val crashlyticsProxy: CrashlyticsProxy, private val notificationsManager: NotificationsManager, private val context: Context) : Consumer<Throwable>, ApiClient {
 
 
     private lateinit var retrofitAdapter: Retrofit
@@ -74,7 +74,7 @@ class ApiClientImpl//private OnHabitsAPIResult mResultListener;
                 .filter { it.data != null }
                 .map { habitResponse ->
                     if (habitResponse.notifications != null) {
-                        popupNotificationsManager.setNotifications(habitResponse.notifications)
+                        notificationsManager.setNotifications(habitResponse.notifications)
                     }
                     habitResponse.data
                 }
@@ -87,7 +87,7 @@ class ApiClientImpl//private OnHabitsAPIResult mResultListener;
     private var lastAPICallURL: String? = null
 
     init {
-        this.popupNotificationsManager.setApiClient(this)
+        this.notificationsManager.setApiClient(this)
 
         HabiticaBaseApplication.component?.inject(this)
         crashlyticsProxy.setUserIdentifier(this.hostConfig.user)
@@ -337,7 +337,7 @@ class ApiClientImpl//private OnHabitsAPIResult mResultListener;
     override fun validateSubscription(request: SubscriptionValidationRequest): Flowable<Any> {
         return apiService.validateSubscription(request).map { habitResponse ->
             if (habitResponse.notifications != null) {
-                popupNotificationsManager.showNotificationDialog(habitResponse.notifications)
+                notificationsManager.showNotificationDialog(habitResponse.notifications)
             }
             habitResponse.getData()
         }
@@ -346,7 +346,7 @@ class ApiClientImpl//private OnHabitsAPIResult mResultListener;
     override fun validateNoRenewSubscription(request: PurchaseValidationRequest): Flowable<Any> {
         return apiService.validateNoRenewSubscription(request).map { habitResponse ->
             if (habitResponse.notifications != null) {
-                popupNotificationsManager.showNotificationDialog(habitResponse.notifications)
+                notificationsManager.showNotificationDialog(habitResponse.notifications)
             }
             habitResponse.getData()
         }
@@ -571,7 +571,7 @@ class ApiClientImpl//private OnHabitsAPIResult mResultListener;
     override fun validatePurchase(request: PurchaseValidationRequest): Flowable<PurchaseValidationResult> {
         return apiService.validatePurchase(request).map { habitResponse ->
             if (habitResponse.notifications != null) {
-                popupNotificationsManager.showNotificationDialog(habitResponse.notifications)
+                notificationsManager.showNotificationDialog(habitResponse.notifications)
             }
             habitResponse.getData()
         }
