+++ Habitica/src/main/java/com/habitrpg/android/habitica/modules/ApiModule.java
@@ -7,7 +7,7 @@ import com.habitrpg.android.habitica.api.HostConfig;
 import com.habitrpg.android.habitica.api.MaintenanceApiService;
 import com.habitrpg.android.habitica.data.ApiClient;
 import com.habitrpg.android.habitica.data.implementation.ApiClientImpl;
-import com.habitrpg.android.habitica.helpers.PopupNotificationsManager;
+import com.habitrpg.android.habitica.helpers.NotificationsManager;
 import com.habitrpg.android.habitica.proxy.CrashlyticsProxy;
 
 import javax.inject.Singleton;
@@ -34,14 +34,14 @@ public class ApiModule {
 
     @Provides
     @Singleton
-    public PopupNotificationsManager providesPopupNotificationsManager(Context context) {
-        return new PopupNotificationsManager(context);
+    public NotificationsManager providesPopupNotificationsManager(Context context) {
+        return new NotificationsManager(context);
     }
 
     @Provides
     @Singleton
-    public ApiClient providesApiHelper(GsonConverterFactory gsonConverter, HostConfig hostConfig, CrashlyticsProxy crashlyticsProxy, PopupNotificationsManager popupNotificationsManager, Context context) {
-        return new ApiClientImpl(gsonConverter, hostConfig, crashlyticsProxy, popupNotificationsManager, context);
+    public ApiClient providesApiHelper(GsonConverterFactory gsonConverter, HostConfig hostConfig, CrashlyticsProxy crashlyticsProxy, NotificationsManager notificationsManager, Context context) {
+        return new ApiClientImpl(gsonConverter, hostConfig, crashlyticsProxy, notificationsManager, context);
     }
 
     @Provides
